The stock span problem
calculate height by using stack : put height in stack, for calculating height compare value of previous index element and if smaller add its height and pop it and  again compare with with pre index elem(by subtracting height of  popped element) do the process until find bigger or stack empty.
OR
we use a stack as an abstract data type to store the days i, h(i), h(h(i)) and so on. When we go from day i-1 to i, we pop the days when the price of the stock was less than or equal to price[i] and then push the value of day i back into the stack.