Evaluation of Postfix Expression 231*+9- is -4
Create an empty stack called operandStack.
Convert the string to a list by using the string method split.
Scan the token list from left to right.
If the token is an operand, convert it from a string to an integer and push the value onto the operandStack.
If the token is an operator, *, /, +, or -, it will need two operands. Pop the operandStack twice. The first pop is the second operand and the second pop is the first operand. Perform the arithmetic operation. Push the result back on the operandStack.
When the input expression has been completely processed, the result is on the stack. Pop the operandStack and return the value.